ref(debugMeta): Convert debugmeta to typescript
