+++ src/sentry/static/sentry/app/components/debugFileFeature.tsx
@@ -24,7 +24,7 @@ const FEATURE_TOOLTIPS = {
 
 type Props = {
   feature: 'symtab' | 'debug' | 'unwind' | 'sources';
-  available?: true;
+  available?: boolean;
 };
 
 const DebugFileFeature = ({available, feature}: Props) => {
@@ -52,6 +52,10 @@ const DebugFileFeature = ({available, feature}: Props) => {
       </Tag>
     </Tooltip>
   );
+};
+
+DebugFileFeature.defaultProps = {
+  available: true,
 };
 
 DebugFileFeature.propTypes = {
