+++ src/sentry/static/sentry/app/components/events/interfaces/threads/threadSelector/filterThreadInfo.tsx
@@ -1,5 +1,5 @@
-import {Thread, Frame} from 'app/types/events';
-import {Event} from 'app/types';
+import {Thread} from 'app/types/events';
+import {Event, Frame} from 'app/types';
 import {trimPackage} from 'app/components/events/interfaces/frame/utils';
 
 import getThreadStacktrace from './getThreadStacktrace';
